+++ resources/ts/app.ts
@@ -7,6 +7,8 @@ import CodeSampleModifier from '@/Modifiers/CodeSampleModifier'
 import LogoModifier from '@/Modifiers/LogoModifier'
 import HomeHeroModifier from '@/Modifiers/HomeHeroModifier'
 import ContributeModifier from '@/Modifiers/ContributeModifier'
+import ToggleCommentsModifier from '@/Modifiers/ToggleCommentsModifier'
+import PlayButtonAdder from '@/Adders/PlayButtonAdder'
 
 new Extension()
     .applyAdders([
@@ -14,10 +16,12 @@ new Extension()
         new HomeLinksAdder(),
         new SearchIconAdder(),
         new CopyButtonAdder(),
+        new PlayButtonAdder(),
     ])
     .applyModifiers([
         new CodeSampleModifier(),
         new HomeHeroModifier(),
         new LogoModifier(),
         new ContributeModifier(),
+        new ToggleCommentsModifier(),
     ])