+++ resources/ts/storage/randomVideos.ts
@@ -2,279 +2,298 @@ import type { RandomVideo } from '@/types'
 
 const randomVideos: RandomVideo[] = [
     {
-        title: "Rasmus Lerdorf – 25 years of PHP",
-        link: "https://youtu.be/Qa_xVjTiOUw",
-        img: "rasmus-lerdorf–25-years-of-php.jpg"
-    },
-    {
         title: "How to get ready for PHP 8 | Sebastian Bergmann",
         link: "https://youtu.be/RjiOKgmqosQ",
-        img: "how-to-get-ready-for-php-8.jpg"
+        img: "how-to-get-ready-for-php-8.jpg",
+        date: "2018-11-27",
     },
     {
         title: "Factory Pattern - Object Oriented PHP Tutorial",
         link: "https://youtu.be/H5lor4wyl1Y",
-        img: "object-oriented-php-tutorial.jpg"
-    },
-    {
-        title: " Introduction To OOP PHP | Object Oriented PHP Tutorial For Beginners | PHP Tutorial | mmtuts",
-        link: "https://youtu.be/Anz0ArcQ5kI",
-        img: "object-oriented-php-mmtuts.jpg"
+        img: "object-oriented-php-tutorial.jpg",
+        date: "2019-12-02",
     },
     {
         title: "Php Objects & Static Methods (Object Oriented Php 2020)",
         link: "https://youtu.be/_Zd9B_jUwlw",
-        img: "objects-and-static-methods.jpg"
+        img: "objects-and-static-methods.jpg",
+        date: "2020-01-20",
     },
     {
         title: "Building a PHP Framework ep1: The dependency injection container",
         link: "https://youtu.be/8bfuR6nEpwY",
-        img: "the-dependency-injection-container.jpg"
+        img: "the-dependency-injection-container.jpg",
+        date: "2018-05-22",
     },
     {
         title: "Marco Pivetta – From helpers to middleware",
         link: "https://youtu.be/MjpiHy_e8kQ",
-        img: "from-helpers-to-middleware.jpg"
+        img: "from-helpers-to-middleware.jpg",
+        date: "2019-12-05",
     },
     {
         title: "PHP UK Conference 2018 - Kat Zien - Performance optimisation: how do I go about it?",
         link: "https://youtu.be/4hB7tGcz1l8",
-        img: "performance-optimisation-how-do-I-go-about-it.jpg"
+        img: "performance-optimisation-how-do-I-go-about-it.jpg",
+        date: "2018-03-29",
     },
     {
         title: "Test to Break Principles - Rob Waller - PHP UK Conference 2019",
         link: "https://youtu.be/JDwFyRXbVHQ",
-        img: "test-to-break-principles-rob-waller.jpg"
-    },
-    {
-        title: "Compiling PHP from source :: Writing tests for PHP source",
-        link: "https://youtu.be/GQRwcCFYhMA",
-        img: "compiling-php-from-source.jpg"
+        img: "test-to-break-principles-rob-waller.jpg",
+        date: "2019-03-20",
     },
     {
         title: "PHP Solid Principles ~ Lesson 1: Single Responsibility",
         link: "https://youtu.be/ARxZV8OZ8Cg",
-        img: "php-solid-principles-single-responsibility.jpg"
+        img: "php-solid-principles-single-responsibility.jpg",
+        date: "2019-12-03",
     },
     {
         title: "Top 100 PHP Functions ( 1 - 10 ) | Learn PHP Programming",
         link: "https://youtu.be/t9FrpTZm1ds",
-        img: "top-100-php-functions-1-10.jpg"
-    },
-    {
-        title: "PHP UK Conference 2017 - David Zuelke - I saw the future, and its web servers were written in PHP",
-        link: "https://youtu.be/VjJBR8YAxoE",
-        img: "david-zuelke-i-saw-the-future-and-its-web.jpg"
+        img: "top-100-php-functions-1-10.jpg",
+        date: "2018-11-19",
     },
     {
         title: "What Is Object-Oriented Programming?",
         link: "https://youtu.be/3rrnnbz-mY8",
-        img: "What-Is-Object-Oriented-Programming.jpg"
+        img: "What-Is-Object-Oriented-Programming.jpg",
+        date: "2018-04-06",
     },
     {
         title: "Controlling the Variables - Thijs Feryn - PHP UK Conference 2019",
         link: "https://youtu.be/f3EU-sd9KPY",
-        img: "controlling-the-variables.jpg"
+        img: "controlling-the-variables.jpg",
+        date: "2019-03-19",
     },
     {
         title: "Create a WordPress Plugin from Scratch - Part 1",
         link: "https://youtu.be/Z7QfH-s-15s",
-        img: "Create-a-WordPress-Plugin-from-Scratch-Part-1.jpg"
-    },
-    {
-        title: "PHP Functions Tutorial - Learn PHP Programming",
-        link: "https://youtu.be/HvxQww-7NGA",
-        img: "php-functions-tutorial.jpg"
+        img: "Create-a-WordPress-Plugin-from-Scratch-Part-1.jpg",
+        date: "2017-08-11",
     },
     {
         title: "PHP cURL Tutorial - Learn PHP Programming",
         link: "https://youtu.be/1X2-UEUqrd8",
-        img: "PHP-cURL-Tutorial-Learn-PHP-Programming.jpg"
+        img: "PHP-cURL-Tutorial-Learn-PHP-Programming.jpg",
+        date: "2016-10-31",
     },
     {
         title: "PHP Regular Expressions Tutorial",
         link: "https://youtu.be/wnu4febXKx0",
-        img: "PHP-Regular-Expressions-Tutorial.jpg"
+        img: "PHP-Regular-Expressions-Tutorial.jpg",
+        date: "2016-11-27",
     },
     {
         title: "Set up Xdebug for Local PHP Development",
         link: "https://youtu.be/iSSjeelN5NU",
-        img: "Set-up-Xdebug-for-Local-PHP-Development.jpg"
+        img: "Set-up-Xdebug-for-Local-PHP-Development.jpg",
+        date: "2017-08-09",
     },
     {
         title: "PHP UK Conference 2018 - Mark Niebergall -Debugging PHP with Xdebug",
         link: "https://youtu.be/4wkHBwkhu7g",
-        img: "PHP-UK-Conference-2018-Mark-Niebergall-Debugging-PHP-with-Xdebug.jpg"
+        img: "PHP-UK-Conference-2018-Mark-Niebergall-Debugging-PHP-with-Xdebug.jpg",
+        date: "2018-03-29",
     },
     {
         title: "Functional Programming in PHP",
         link: "https://youtu.be/LZh4_q04aKo",
-        img: "Functional-Programming-in-PHP.jpg"
-    },
-    {
-        title: "PHP Standard Library Part 1: Data structures",
-        link: "https://youtu.be/6UkejPv3ck8",
-        img: "PHP-Standard-Library-Part-1-Data-structures.jpg"
-    },
-    {
-        title: "Beautiful Slim CSRF Protection: Introduction (1/4)",
-        link: "https://youtu.be/GHa2XBAx7WU",
-        img: "Beautiful-Slim-CSRF-Protection-Introduction-1-4.jpg"
+        img: "Functional-Programming-in-PHP.jpg",
+        date: "2018-10-05",
     },
     {
         title: "PHP REST API From Scratch [1] - Database & Read",
         link: "https://youtu.be/OEWXbpUMODk",
-        img: "PHP-REST-API-From-Scratch-1-Database-Read.jpg"
-    },
-    {
-        title: "Using REST APIs in a web application | Quick PHP Tutorial",
-        link: "https://youtu.be/RTjd1nwvlj4",
-        img: "Using-REST-APIs-in-a-web-application-Quick-PHP-Tutorial.jpg"
+        img: "PHP-REST-API-From-Scratch-1-Database-Read.jpg",
+        date: "2018-05-27",
     },
     {
         title: "Authentication With Slim 4 (Lesson 1: Setup Slim)",
         link: "https://youtu.be/3Hg2WPwDyG8",
-        img: "Authentication-With-Slim-4-Lesson-1-Setup-Slim.jpg"
-    },
-    {
-        title: "PHP web development environment with Docker tutorial - 001",
-        link: "https://youtu.be/2ESAi2vq-80",
-        img: "PHP-web-development-environment-with-Docker-tutorial-001.jpg"
+        img: "Authentication-With-Slim-4-Lesson-1-Setup-Slim.jpg",
+        date: "2020-02-17",
     },
     {
         title: "PHP UK Conference 2017 - Gareth Ellis - Introduction to SOLID",
         link: "https://youtu.be/86Tt2pW9pv4",
-        img: "PHP-UK-Conference-2017-Gareth Ellis-Introduction-to-SOLID.jpg"
-    },
-    {
-        title: "PHP UK Conference 2017 - Anna Filina - Unit Testing by Example",
-        link: "https://youtu.be/ESl-ncXA4G0",
-        img: "PHP-UK-Conference-2017-Anna-Filina-Unit-Testing-by-Example.jpg"
-    },
-    {
-        title: "18: Apply The MVC Model Using OOP PHP | MVC Model Tutorial | Object Oriented PHP Tutorial",
-        link: "https://youtu.be/Fg0CP-ri87U",
-        img: "Apply-The-MVC-Model-Using-OOP-PHP.jpg"
+        img: "PHP-UK-Conference-2017-Gareth Ellis-Introduction-to-SOLID.jpg",
+        date: "2017-04-03",
     },
     {
         title: "#devbreak19 - 25 Years of PHP",
         link: "https://youtu.be/80ysU2Xl4Zk",
-        img: "devbreak19-25-Years-of-PHP.jpg"
+        img: "devbreak19-25-Years-of-PHP.jpg",
+        date: "2019-08-07",
     },
     {
         title: "25 years of PHP - Rasmus Lerdorf",
         link: "https://youtu.be/nmD1Q4FsXCc",
-        img: "25-years-of-PHP-Rasmus-Lerdorf-video.jpg"
+        img: "25-years-of-PHP-Rasmus-Lerdorf-video.jpg",
+        date: "2019-07-11",
     },
     {
         title: "PHP IS BACK IN 2020. 3 Myths and Which Big Companies Are Using PHP",
         link: "https://youtu.be/FaufgEN3H2g",
-        img: "PHP-IS-BACK-IN-2020.jpg"
-    },
-    {
-        title: "PhpStorm setup for PHP web development with Docker - 002",
-        link: "https://youtu.be/FQWL-JTZvf4",
-        img: "PhpStorm-setup-for-PHP-web-development-with-Docker-002.jpg"
+        img: "PHP-IS-BACK-IN-2020.jpg",
+        date: "2020-01-06",
     },
     {
         title: "What is Class in PHP - OOP Tutorial - in 5 Minutes",
         link: "https://youtu.be/pObVzQjRWtg",
-        img: "What-is-Class-in-PHP-OOP-Tutorial-in-5-Minutes.jpg"
+        img: "What-is-Class-in-PHP-OOP-Tutorial-in-5-Minutes.jpg",
+        date: "2020-04-03",
     },
     {
         title: "Object Oriented programming in PHP | PHP for Beginners - Part 20",
         link: "https://youtu.be/cNe_24n3maY",
-        img: "Object-Oriented-programming-in-PHP.jpg"
+        img: "Object-Oriented-programming-in-PHP.jpg",
+        date: "2020-04-16",
     },
     {
         title: "Serverless PHP applications with Bref - Matthiew Napoli - PHP UK 2020",
         link: "https://youtu.be/R2V4QTM2aes",
-        img: "Serverless-PHP-applications-with-Bre-Matthiew-Napoli-PHP-UK-2020.jpg"
+        img: "Serverless-PHP-applications-with-Bre-Matthiew-Napoli-PHP-UK-2020.jpg",
+        date: "2020-03-12",
     },
     {
         title: "What's new in PHP 8.0",
         link: "https://youtu.be/uU1-ZqIbYes",
-        img: "Whats-new-in-PHP-8-0.webp"
+        img: "Whats-new-in-PHP-8-0.webp",
+        date: "2020-11-28",
     },
     {
         title: "Laravel Crash Course 2020",
         link: "https://youtu.be/MFh0Fd7BsjE",
-        img: "Laravel-Crash-Course-2020.webp"
+        img: "Laravel-Crash-Course-2020.webp",
+        date: "2020-11-16",
     },
     {
         title: "5 Exciting New Features in PHP 8.1",
         link: "https://youtu.be/T8QZkpVvDxc",
-        img: "5-Exciting-New-Features-in-PHP-81.webp"
+        img: "5-Exciting-New-Features-in-PHP-81.webp",
+        date: "2021-10-12",
     },
     {
         title: "What’s New in PHP 8.1: Enums, First-Class Callables, Fibers, Readonly Properties, and More",
         link: "https://youtu.be/rgrQSmUiFJQ",
-        img: "Whats-New-in-PHP-81-Enums-First-Class-Callables-Fibers-Readonly-Properties-and-More.webp"
+        img: "Whats-New-in-PHP-81-Enums-First-Class-Callables-Fibers-Readonly-Properties-and-More.webp",
+        date: "2021-11-25",
     },
     {
         title: "Unit Testing with PHP Unit",
         link: "https://youtu.be/a5ZKCFINUkU",
-        img: "Unit-Testing-with-PHP-Unit.webp"
+        img: "Unit-Testing-with-PHP-Unit.webp",
+        date: "2021-03-07",
     },
     {
         title: "PHP Unit Testing with PHPUnit | Automated PHP Testing Tutorial [2021]",
         link: "https://youtu.be/kkU43JdJQBE",
-        img: "PHP-Unit-Testing-with-PHPUnit-Automated-PHP-Testing-Tutorial-2021.webp"
+        img: "PHP-Unit-Testing-with-PHPUnit-Automated-PHP-Testing-Tutorial-2021.webp",
+        date: "2021-05-10",
     },
     {
         title: "5 Reasons to Use PHP 8: New Syntax Features",
         link: "https://youtu.be/F-AWT88vsyM",
-        img: "5-Reasons-to-Use-PHP-8-New-Syntax-Features.webp"
+        img: "5-Reasons-to-Use-PHP-8-New-Syntax-Features.webp",
+        date: "2021-11-05",
     },
     {
         title: "The Future of PHP",
         link: "https://www.youtube.com/watch?v=wuJ2W3n0vXg",
-        img: "The-Future-of-PHP.webp"
+        img: "The-Future-of-PHP.webp",
+        date: "2022-08-04",
     },
     {
         title: "Why use PHP in 2022?",
         link: "https://www.youtube.com/watch?v=X5lxnlcjrWM",
-        img: "Why-use-PHP-in-2022.webp"
+        img: "Why-use-PHP-in-2022.webp",
+        date: "2022-06-15",
     },
     {
         title: "The evolution of PHP",
         link: "https://www.youtube.com/watch?v=x9bSUo6TGgY",
-        img: "From-PHP-5-6-to-PHP-8-2.webp"
+        img: "From-PHP-5-6-to-PHP-8-2.webp",
+        date: "2022-07-17",
     },
     {
         title: "Celebrating PHP 8.3",
         link: "https://www.youtube.com/live/VWryF035B6U?si=HzqFkcaMdIBgGlVa",
         img: "Celebrating-PHP-8-3.webp",
+        date: "2023-11-23",
     },
     {
         title: "Breaking changes in PHP 8.3!",
         link: "https://youtu.be/mGf_ShkSdmE?si=EBC9q0RDM5gIvLBR",
         img: "Breaking-changes-in-PHP-8-3.webp",
+        date: "2023-10-26",
     },
     {
         title: "PHP For Beginners - Complete Laracasts Course",
         link: "https://youtu.be/fw5ObX8P6as?si=L7E6jf4UNTNOxLy1",
         img: "PHP-For-Beginners-Complete-Laracasts-Course.webp",
+        date: "2023-11-21",
     },
     {
         title: "In Defence of PHP - Stephen Rees-Carter - NDC Security 2023",
         link: "https://youtu.be/zlPBufBzscg?si=7yYQlVjPl5A-ej7N",
         img: "In-Defence-of-PHP-Stephen-Rees-Carter-NDC-Security-2023.webp",
+        date: "2023-04-20",
     },
     {
         title: "Readonly clones, #[Override], and json_validate: what's new in PHP 8.3",
         link: "https://youtu.be/nJFsD0bnlTI?si=Y51fxIQFM3C91CTZ",
         img: "Readonly-clones-and-json_validate-whats-new-in-PHP-8-3.webp",
+        date: "2023-10-05",
     },
     {
-        title: "\"Developing Symfony applications in 2023\" [eng] / Nicolas Grekas",
+        title: `"Developing Symfony applications in 2023" [eng] / Nicolas Grekas`,
         link: "https://youtu.be/EwAVP3BoNOg?si=zTioUaqULIyZsMdH",
-        img: "Developing-Symfony-applications-in-2023-Nicolas-Grekas.webp"
+        img: "Developing-Symfony-applications-in-2023-Nicolas-Grekas.webp",
+        date: "2023-11-24",
     },
     {
         title: "Is PHP 8.3 adding types to class constants?!",
         link: "https://youtu.be/E8lbWznz1d0?si=OWZMOXmR-MmIvjo5",
-        img: "Is-PHP-8-3-adding-types-to-class-constants.webp"
+        img: "Is-PHP-8-3-adding-types-to-class-constants.webp",
+        date: "2023-10-13",
+    },
+    {
+        title: "Clean your controller with Laravel prepareForValidation method",
+        link: "https://youtu.be/RbMlMAULJm8?si=lj_ExcBmAeKgIkMs",
+        img: "Clean your controller with Laravel prepareForValidation method.webp",
+        date: "2022-09-24",
+    },
+    {
+        title: "Logging Unique Views in Laravel with Redis: Playing with Redis sets and HyperLogLog (1/4)",
+        link: "https://youtu.be/FGLDa3e4ZGM?si=Z4_l2wqn09oPViiI",
+        img: "Logging Unique Views in Laravel with Redis- Playing with Redis sets and HyperLogLog (1-4).webp",
+        date: "2024-02-20",
+    },
+    {
+        title: "Up And Running with Pest: Introduction and demo (01/35)",
+        link: "https://youtu.be/jyfoHw05Mbw?si=8VfLZqiwmJPf1TQ6",
+        img: "Up And Running with Pest- Introduction and demo (01-35).webp",
+        date: "2022-03-28",
+    },
+    {
+        title: "What's New in PHP 8.3 - Derick Rethans",
+        link: "https://youtu.be/3U0DGhzSH2U?si=x6EHaR5jvJmorgV0",
+        img: "What's New in PHP 8.3 - Derick Rethans.webp",
+        date: "2024-02-22",
+    },
+    {
+        title: "Boosting Frontend Speed: Quick Wins for Backend Developers - Paul Conroy",
+        link: "https://youtu.be/BZuWGil_7kY?si=zqeyIP2afbIrT2ts",
+        img: "Boosting Frontend Speed- Quick Wins for Backend Developers - Paul Conroy.webp",
+        date: "2024-02-21",
+    },
+    {
+        title: "PHP is the future",
+        link: "https://youtu.be/xmvD_EjNE-4?si=iu_8S-Imzwd-jPqf",
+        img: "PHP is the future.webp",
+        date: "2023-05-12",
     },
 ]